Split file handling class into input and output specific classes
Refactor SAVE_FILE_PATH from SaveFileHandler to StringStorage
Split SaveFileHandler class into classes InputFileHandler and OutputFileHandler
